# public/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u638c\u4e0a\u5357\u660cAPI v1",
        "termsOfService": "",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/members/login": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u4f1a\u5458\u767b\u5f55",
                "parameters": [
                    {
                        "name": "member_name",
                        "in": "query",
                        "description": "\u4f1a\u5458\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u767b\u5f55\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
        "/members/exlogin": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u7b2c\u4e09\u65b9\u767b\u5f55",
                "parameters": [
                    {
                        "name": "member_name",
                        "in": "query",
                        "description": "\u4f1a\u5458\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "nick_name",
                        "in": "query",
                        "description": "\u6635\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "avatar_url",
                        "in": "query",
                        "description": "\u5934\u50cfURL",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "\u6765\u6e90(qq,wx,wb)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u767b\u5f55\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
        "/members/register": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u4f1a\u5458\u6ce8\u518c",
                "parameters": [
                    {
                        "name": "member_name",
                        "in": "query",
                        "description": "\u4f1a\u5458\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "captcha",
                        "in": "query",
                        "description": "\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6ce8\u518c\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
        "/members/info": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u83b7\u53d6\u4f1a\u5458\u4fe1\u606f",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "description": "\u7ad9\u70b9ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6b63\u5e38"
                    },
                    "401": {
                        "description": "\u65e0\u6548"
                    }
                }
            }
        },
        "/members/avatar": {
            "post": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u4fee\u6539\u5934\u50cf",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "avatar_file",
                        "in": "formData",
                        "description": "\u56fe\u7247\u6587\u4ef6",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u4e0a\u4f20\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230"
                    }
                }
            }
        },
        "/members/info/nick": {
            "post": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u4fee\u6539\u4f1a\u5458\u6635\u79f0",
                "parameters": [
                    {
                        "name": "nick_name",
                        "in": "query",
                        "description": "\u6635\u79f0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6b63\u5e38"
                    },
                    "401": {
                        "description": "\u65e0\u6548"
                    }
                }
            }
        },
        "/members/token/status": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u83b7\u53d6token\u72b6\u6001",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6b63\u5e38"
                    },
                    "401": {
                        "description": "\u65e0\u6548"
                    }
                }
            }
        },
        "/members/mobile/captcha": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u83b7\u53d6\u9a8c\u8bc1\u7801",
                "parameters": [
                    {
                        "name": "site_id",
                        "in": "query",
                        "description": "\u7ad9\u70b9ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "\u7c7b\u578b(0:\u627e\u56de\u5bc6\u7801,1:\u6ce8\u518c,2:\u91cd\u7f6e\u5bc6\u7801,3:\u7ed1\u5b9a\u624b\u673a,4:\u89e3\u9664\u7ed1\u5b9a\u624b\u673a,5:\u767b\u5f55\u9a8c\u8bc1)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6b63\u5e38"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
        "/members/mobile/bind": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u7ed1\u5b9a\u624b\u673a\u53f7",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "\u624b\u673a\u53f7",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "captcha",
                        "in": "query",
                        "description": "\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u7ed1\u5b9a\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
        "/members/mobile/unbind": {
            "get": {
                "tags": [
                    "/members \u4f1a\u5458"
                ],
                "summary": "\u89e3\u7ed1\u624b\u673a\u53f7",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "captcha",
                        "in": "query",
                        "description": "\u9a8c\u8bc1\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u89e3\u7ed1\u6210\u529f"
                    },
                    "404": {
                        "description": "\u6ca1\u6709\u627e\u5230\u8def\u7531"
                    }
                }
            }
        },
...
    },
...
}